refactor[openbeats]: move BEATs into espnet2/beats/
Consolidate the BEATs encoder, tokenizer, random projection
quantizer, and shared utils into a single espnet2/beats/ package
(previously split between espnet2/asr/encoder/ and
espnet2/speechlm/tokenizer/). Test files relocated to
test/espnet2/beats/ to mirror the new layout. Per Shinji's PR #6436
review comments.
