+++ espnet2/tasks/cls.py
@@ -8,7 +8,6 @@ import torch
 from typeguard import typechecked
 
 from espnet2.asr.encoder.abs_encoder import AbsEncoder
-from espnet2.asr.encoder.beats_encoder import BeatsEncoder
 from espnet2.asr.encoder.conformer_encoder import ConformerEncoder
 from espnet2.asr.encoder.transformer_encoder import TransformerEncoder
 from espnet2.asr.frontend.abs_frontend import AbsFrontend
@@ -21,6 +20,7 @@ from espnet2.asr.preencoder.linear import LinearProjection
 from espnet2.asr.preencoder.sinc import LightweightSincConvs
 from espnet2.asr.specaug.abs_specaug import AbsSpecAug
 from espnet2.asr.specaug.specaug import SpecAug
+from espnet2.beats.encoder import BeatsEncoder
 from espnet2.cls.decoder.abs_decoder import AbsDecoder
 from espnet2.cls.decoder.linear_decoder import LinearDecoder
 from espnet2.cls.espnet_model import ESPnetClassificationModel
