+++ test/espnet2/beats/test_utils.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from espnet2.speechlm.tokenizer.beats_utils import (
+from espnet2.beats.utils import (
     beats_frontend,
     forward_padding_mask_conv,
     freeze_conv_module,
